+++ app/modules/custom_reports/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 
 from .routes import custom_reports_bp
 from .service import CustomReportsService